ai-engine/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
torch
torchvision
transformers
opencv-python
Pillow
numpy
scipy
pydantic<2.0
python-dotenv==1.0.0
aiofiles==23.2.1
pytest
pytest-asyncio
httpx
